Restructure breakpoints definitions into their module
